docs(sessions): explain trim safety in an info bubble and on the website
The trim editor gains an InfoHint (extracted from BookDetailPage into a
shared component) stating what a trim does and does not touch, and that
the previous duration survives in the audit log. Website docs grow a
'Fixing a runaway session' section on the stats page and an 'Idle time
cap' section on the KOReader page.
